pyproject.toml
[tool.poetry]
name = "sentiment-analysis"
version = "0.1.0"
description = "Sentiment analysis on market comments"
authors = ["Zernyshkina Ekaterina <zernyshkina.ea@phystech.edu>"]
packages = [{include = "sentiment_analysis"}]

[tool.poetry.dependencies]
python = "^3.10"
tensorflow = "^2.15.0"
dvc = "^3.50.1"
hydra-core = "^1.3.2"
mlflow = "^2.14.2"
nltk = "^3.8.1"
scikit-learn = "^1.4.2"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
onnx = "^1.16.0"
onnxruntime = "^1.17.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
prettier = "^0.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
line_length = 119
ignore = ["E203", "W503"]
